requirements.txt
# Core dependencies
fastf1>=3.0.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
kaleido>=0.2.1  # For plotly image export

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.10.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver
alembic>=1.12.0  # Database migrations

# GUI
PySide6>=6.5.0

# CLI
questionary>=2.0.0
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
